extracted handleBrowserFeaturesMethod
Because I am trying to figure out how to do some IO on the server side, and was hoping ghci would give me a useful type signature (e.g. print or
log the query, save the browser property in the database. I eventually want to use a record with more properties as parameter for the command,
but am going for a vertical slice first).

Figured out part of the type signature with IO in it by trying some IO, leaving off
the type signature and reading the error messages.