WIP introduce Sum type for supported browser features
Idea is to pair checker function with a BrowserSupportValue, and get a
list of all supported Browser Values back.

Some conditional compilation, it seems catMaybes is useful, but it does
not exist in Fay, because it has a limited monad library.

The synonym value allValues exists to get possible compile errors on
either side.