WIP Html shadow DOM from Algebraic Data Type
Inspired by http://www.cs.york.ac.uk/fp/HaXml/icfp99.html

Reason for doing it like this, is that we can keep all UI generating
code pure, and that we can generate HTML on the server side, and DOM
objects in the browser, by just walking a tree with a few functions.

Constructing the tree is a bit verbose, we might make some helper
functions for the most used combinations (e.g. h1, h2, p, ahref for an A
tag with just href attribute and some text, ul
[list-ofthings-to-go-in-lis] . I tend to use just a few from the many
html elements. The blaze-html source code has complete sets for all of
the attributes and elements, in case we want to go for full-blown
typing.

The half-typed half-stringly-typed structure here makes it easy to
generate HTML tags and DOM elements with a small amount of functions,
and should make it easy / clean enough to construct html in haskell.