Spike - use of Text instead of string
Text should be represented as regular javascript arrays, and is
apparently safer to use when marshalling in requests.

Figuring out how to get it work, was, well, work. It's in a separata
module TextPeriment for now, because Text is defined in a couple of
places (amongst others, in Language.Fay.Yesod ), so converting all
String to Text in Home.hs was a bit too much work.

For some reason Fay.Text only implements fromString, not toString
(toString converts a String to a Text implicitly when OverloadedStrings
and RebindableSyntax language extensions are enabled).